// src/app/Models/RecordModel.ts
export interface RecordModel {
    timeStamp: string;
    EventNum: string;
    EventCategory: string;
    EventText: string;
    nVal?: string;
    qVal?: string;
    tVal?: string;
    dVal?: string;
    uVal?: string;
    cVal?: string;
    rVal?: string;
    LocationNum: string;
    Latitude: string;
    Longitude: string;
    State: string;
    PoliceStation: string;
    Special: string;
    id: string;
}